+++ test/int/context.js
@@ -1,5 +1,6 @@
 const VaultPKIEngine = require('../../src/pki_engine/VaultPKIEngine');
 const Constants = require('../../src/constants/Constants');
+const { logger } = require('../../src/log/logger');
 
 exports.createContext = async () => {
   const pkiEngine = new VaultPKIEngine({
@@ -7,8 +8,8 @@ exports.createContext = async () => {
     signExpiryHours: 1,
   });
   await pkiEngine.connect()
-    .then(() => { console.log('VAULT is CONNECTED!'); })
-    .catch((err) => { console.error(`VAULT is NOT CONNECTED: ${err?.message}`, err); });
+    .then(() => { logger.debug('VAULT is CONNECTED!'); })
+    .catch((err) => { logger.error(`VAULT is NOT CONNECTED: ${err?.message}`, err); });
   return {
     pkiEngine,
   };