+++ test/int/test-env-setup.js
@@ -11,7 +11,7 @@ process.env.LOG_LEVEL = 'warn';
 
 // Server
 process.env.PORT = '3001';
-process.env.CLIENT_URL = 'http://localhost:3000/';
+process.env.CLIENT_URL = 'http://mcm.localhost/';
 
 // Database
 process.env.DATABASE_HOST = 'localhost';
@@ -21,7 +21,7 @@ process.env.DATABASE_PASSWORD = 'mcm';
 process.env.DATABASE_SCHEMA = 'mcm';
 
 // Vault PKI
-process.env.VAULT_ENDPOINT = 'http://localhost:8233';
+process.env.VAULT_ENDPOINT = 'http://vault.mcm.localhost';
 process.env.VAULT_AUTH_METHOD = 'APP_ROLE';
 process.env.VAULT_ROLE_ID_FILE = './docker/vault/tmp/role-id';
 process.env.VAULT_ROLE_SECRET_ID_FILE = './docker/vault/tmp/secret-id';
@@ -36,8 +36,8 @@ process.env.AUTH_ENABLED = 'true';
 
 // Keycloak integration
 process.env.KEYCLOAK_ENABLED = 'true';
-process.env.KEYCLOAK_BASE_URL = 'http://localhost:8080';
-process.env.KEYCLOAK_DISCOVERY_URL = 'http://localhost:8080/realms/dfsps/.well-known/openid-configuration';
+process.env.KEYCLOAK_BASE_URL = 'http://keycloak.mcm.localhost';
+process.env.KEYCLOAK_DISCOVERY_URL = 'http://keycloak.mcm.localhost/realms/dfsps/.well-known/openid-configuration';
 process.env.KEYCLOAK_ADMIN_CLIENT_ID = 'connection-manager-api-service';
 process.env.KEYCLOAK_ADMIN_CLIENT_SECRET = 'dfsps123';
 process.env.KEYCLOAK_DFSPS_REALM = 'dfsps';
@@ -47,10 +47,12 @@ process.env.KEYCLOAK_AUTO_CREATE_ACCOUNTS = 'true';
 process.env.OPENID_ENABLED = 'true';
 process.env.OPENID_ENABLE_2FA = 'true';
 process.env.OPENID_ALLOW_INSECURE = 'true';
-process.env.OPENID_DISCOVERY_URL = 'http://localhost:8080/realms/dfsps/.well-known/openid-configuration';
+process.env.OPENID_DISCOVERY_URL = 'http://keycloak.mcm.localhost/realms/dfsps/.well-known/openid-configuration';
 process.env.OPENID_CLIENT_ID = 'connection-manager-auth-client';
 process.env.OPENID_CLIENT_SECRET = 'dfsps456';
-process.env.OPENID_REDIRECT_URI = 'http://localhost:3001/api/auth/callback';
+process.env.OPENID_REDIRECT_URI = 'http://mcm.localhost/api/auth/callback';
+
+// Cookie and role settings
 process.env.OPENID_JWT_COOKIE_NAME = 'MCM-API_ACCESS_TOKEN';
 process.env.OPENID_EVERYONE_ROLE = 'everyone';
 process.env.OPENID_MTA_ROLE = 'mta';
@@ -61,4 +63,3 @@ process.env.SESSION_SECRET = 'test-session-secret-for-integration-tests';
 
 // Features
 process.env.DFSP_WATCHER_ENABLED = 'false';
-